moved file holder to models
